+++ frontend/v1/domain/usecases/dataset-setting/get-dataset-settings-use-case.ts
@@ -5,6 +5,7 @@ import { Question } from "../../entities/question/Question";
 import { Vector } from "../../entities/vector/Vector";
 import { IDatasetRepository } from "../../services/IDatasetRepository";
 import { IDatasetSettingStorage } from "../../services/IDatasetSettingStorage";
+import { RoleService } from "../../services/RoleService";
 import {
   FieldRepository,
   MetadataRepository,
@@ -14,6 +15,7 @@ import {
 
 export class GetDatasetSettingsUseCase {
   constructor(
+    private readonly roleService: RoleService,
     private readonly datasetRepository: IDatasetRepository,
     private readonly questionRepository: QuestionRepository,
     private readonly fieldRepository: FieldRepository,
@@ -23,6 +25,28 @@ export class GetDatasetSettingsUseCase {
   ) {}
 
   async execute(datasetId: string): Promise<DatasetSetting> {
+    const isAdminOrRole = this.roleService.isAdminOrOwner();
+
+    const datasetSetting = isAdminOrRole
+      ? await this.createDatasetSettingsForAdminOrOwner(datasetId)
+      : await this.createDatasetSettingsForAnnotator(datasetId);
+
+    this.datasetSettingStorage.save(datasetSetting);
+
+    return datasetSetting;
+  }
+
+  private async createDatasetSettingsForAnnotator(
+    datasetId: string
+  ): Promise<DatasetSetting> {
+    const dataset = await this.datasetRepository.getById(datasetId);
+
+    return new DatasetSetting(dataset);
+  }
+
+  private async createDatasetSettingsForAdminOrOwner(
+    datasetId: string
+  ): Promise<DatasetSetting> {
     const getDataset = this.datasetRepository.getById(datasetId);
     const getQuestions = this.questionRepository.getQuestions(datasetId);
     const getFields = this.fieldRepository.getFields(datasetId);
@@ -87,16 +111,12 @@ export class GetDatasetSettingsUseCase {
       );
     });
 
-    const datasetSetting = new DatasetSetting(
+    return new DatasetSetting(
       dataset,
       questions,
       fields,
       vectors,
       metadataProperties
     );
-
-    this.datasetSettingStorage.save(datasetSetting);
-
-    return datasetSetting;
   }
 }