+++ src/argilla/client/feedback/dataset/local/mixins.py
@@ -286,11 +286,15 @@ class ArgillaMixin:
         questions = ArgillaMixin.__get_questions(client=httpx_client, id=argilla_id)
         question_name_to_id = {question.name: question.id for question in questions}
 
-        metadata_properties = ArgillaMixin.__add_metadata_properties(
-            metadata_properties=self.metadata_properties, client=httpx_client, id=argilla_id
-        )
+        if self.metadata_properties:
+            ArgillaMixin.__add_metadata_properties(
+                metadata_properties=self.metadata_properties, client=httpx_client, id=argilla_id
+            )
 
-        ArgillaMixin.__add_vectors_settings(vectors_settings=self.vector_settings, client=httpx_client, id=argilla_id)
+        if self.vectors_settings:
+            ArgillaMixin.__add_vectors_settings(
+                vectors_settings=self.vectors_settings, client=httpx_client, id=argilla_id
+            )
 
         ArgillaMixin.__publish_dataset(client=httpx_client, id=argilla_id)
 
@@ -311,7 +315,6 @@ class ArgillaMixin:
             updated_at=new_dataset.updated_at,
             fields=fields,
             questions=questions,
-            metadata_properties=metadata_properties,
             guidelines=self.guidelines,
             allow_extra_metadata=self.allow_extra_metadata,
         )
@@ -385,7 +388,6 @@ class ArgillaMixin:
 
         fields = ArgillaMixin.__get_fields(client=httpx_client, id=existing_dataset.id)
         questions = ArgillaMixin.__get_questions(client=httpx_client, id=existing_dataset.id)
-        metadata_properties = ArgillaMixin.__get_metadata_properties(client=httpx_client, id=existing_dataset.id)
 
         return RemoteFeedbackDataset(
             client=httpx_client,
@@ -396,7 +398,6 @@ class ArgillaMixin:
             updated_at=existing_dataset.updated_at,
             fields=fields,
             questions=questions,
-            metadata_properties=metadata_properties,
             guidelines=existing_dataset.guidelines or None,
             allow_extra_metadata=existing_dataset.allow_extra_metadata,
         )
@@ -493,7 +494,7 @@ class TaskTemplateMixin:
         labels: List[str],
         multi_label: bool = False,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -539,7 +540,7 @@ class TaskTemplateMixin:
     def for_question_answering(
         cls: Type["FeedbackDataset"],
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -575,7 +576,7 @@ class TaskTemplateMixin:
     def for_summarization(
         cls: Type["FeedbackDataset"],
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -605,7 +606,7 @@ class TaskTemplateMixin:
     def for_translation(
         cls: Type["FeedbackDataset"],
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -634,7 +635,7 @@ class TaskTemplateMixin:
         cls: Type["FeedbackDataset"],
         rating_scale: int = 7,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -671,7 +672,7 @@ class TaskTemplateMixin:
         cls: Type["FeedbackDataset"],
         labels: Optional[List[str]] = None,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -704,7 +705,7 @@ class TaskTemplateMixin:
         cls: Type["FeedbackDataset"],
         context: bool = False,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -746,7 +747,7 @@ class TaskTemplateMixin:
         number_of_responses: int = 2,
         context: bool = False,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -804,7 +805,7 @@ class TaskTemplateMixin:
         rating_scale: int = 7,
         context: bool = False,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -844,7 +845,7 @@ class TaskTemplateMixin:
         number_of_responses: int = 2,
         context: bool = False,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """
@@ -874,7 +875,7 @@ class TaskTemplateMixin:
         number_of_retrievals: int = 1,
         rating_scale: int = 7,
         use_markdown: bool = False,
-        guidelines: str = None,
+        guidelines: Optional[str] = None,
         metadata_properties: List[AllowedMetadataPropertyTypes] = None,
     ) -> "FeedbackDataset":
         """