+++ tests/integration/client/feedback/dataset/remote/test_dataset.py
@@ -67,16 +67,7 @@ if TYPE_CHECKING:
 
 
 @pytest.fixture()
-def test_dataset_with_metadata_properties():
-    dataset = FeedbackDataset(
-        fields=[TextField(name="text"), TextField(name="optional", required=False)],
-        questions=[TextQuestion(name="question")],
-    )
-    return dataset
-
-
-@pytest.fixture()
-def test_dataset_with_metadata_properties():
+def test_dataset_with_metadata_properties() -> FeedbackDataset:
     dataset = FeedbackDataset(
         fields=[TextField(name="text"), TextField(name="optional", required=False)],
         questions=[TextQuestion(name="question")],
@@ -92,7 +83,7 @@ def test_dataset_with_metadata_properties():
 @pytest.fixture
 def feedback_dataset() -> FeedbackDataset:
     return FeedbackDataset(
-        fields=[TextField(name="text"), TextField(name="text-2")],
+        fields=[TextField(name="text"), TextField(name="text-2", required=False)],
         questions=[
             TextQuestion(name="text"),
             LabelQuestion(name="label", labels=["label-1", "label-2", "label-3"], required=False),
@@ -105,8 +96,9 @@ def feedback_dataset() -> FeedbackDataset:
             IntegerMetadataProperty(name="integer-metadata", min=0, max=10),
             FloatMetadataProperty(name="float-metadata", min=0, max=10),
         ],
+        vectors_settings=[VectorSettings(name="vector-1", dimensions=3), VectorSettings(name="vector-2", dimensions=4)],
         guidelines="unit test guidelines",
-        allow_extra_metadata=False,
+        allow_extra_metadata=True,
     )
 
 
@@ -116,18 +108,19 @@ class TestRemoteFeedbackDataset:
         "record",
         [
             FeedbackRecord(fields={"text": "Hello world!"}, metadata={}),
-            FeedbackRecord(fields={"text": "Hello world!", "optional": "Bye world!"}, metadata={}),
+            FeedbackRecord(fields={"text": "Hello world!", "text-2": "Bye world!"}, metadata={}),
             FeedbackRecord(fields={"text": "Hello world!"}, metadata={"terms-metadata": "a"}),
             FeedbackRecord(fields={"text": "Hello world!"}, metadata={"unrelated-metadata": "unrelated-value"}),
+            FeedbackRecord(
+                fields={"text": "Hello world!"}, vectors={"vector-1": [1.0, 2.0, 3.0], "vector-2": [1.0, 2.0, 3.0, 4.0]}
+            ),
         ],
     )
-    async def test_add_records(
-        self, owner: "User", test_dataset_with_metadata_properties: FeedbackDataset, record: FeedbackRecord
-    ) -> None:
+    async def test_add_records(self, owner: "User", feedback_dataset: FeedbackDataset, record: FeedbackRecord) -> None:
         api.init(api_key=owner.api_key)
         ws = Workspace.create(name="test-workspace")
 
-        remote = test_dataset_with_metadata_properties.push_to_argilla(name="test_dataset", workspace=ws)
+        remote = feedback_dataset.push_to_argilla(name="test_dataset", workspace=ws)
 
         remote_dataset = FeedbackDataset.from_argilla(id=remote.id)
         remote_dataset.add_records([record])
@@ -257,7 +250,7 @@ class TestRemoteFeedbackDataset:
         assert remote.name == "unit-test-dataset"
         assert remote.workspace.name == workspace.name
         assert remote.guidelines == "unit test guidelines"
-        assert remote.allow_extra_metadata is False
+        assert remote.allow_extra_metadata is feedback_dataset.allow_extra_metadata
 
         for remote_field, field in zip(remote.fields, feedback_dataset.fields):
             assert field.name == remote_field.name
@@ -280,7 +273,7 @@ class TestRemoteFeedbackDataset:
         assert remote.name == "unit-test-dataset"
         assert remote.workspace.name == workspace.name
         assert remote.guidelines == "unit test guidelines"
-        assert remote.allow_extra_metadata is False
+        assert remote.allow_extra_metadata is feedback_dataset.allow_extra_metadata
 
         for remote_field, field in zip(remote.fields, feedback_dataset.fields):
             assert field.name == remote_field.name
@@ -433,14 +426,16 @@ class TestRemoteFeedbackDataset:
         remote_dataset.delete_metadata_properties(names)
         assert len(remote_dataset.metadata_properties) == 0
 
-    def test_adding_vector_settings(self, owner: "User", feedback_dataset: FeedbackDataset):
+    def test_add_vector_settings(self, owner: "User"):
         api.init(api_key=owner.api_key)
         workspace = Workspace.create(name="test-workspace")
+
+        feedback_dataset = FeedbackDataset(fields=[TextField(name="text")], questions=[TextQuestion(name="text")])
 
         remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
         vector_settings = remote_dataset.add_vector_settings(VectorSettings(name="vector", dimensions=10))
 
-        assert len(remote_dataset.vector_settings) == 1
+        assert len(remote_dataset.vectors_settings) == 1
 
         remote_vector_settings = remote_dataset.vector_settings_by_name("vector")
         assert vector_settings.name == remote_vector_settings.name
@@ -449,26 +444,126 @@ class TestRemoteFeedbackDataset:
 
         other_vector_settings = remote_dataset.add_vector_settings(VectorSettings(name="other-vector", dimensions=100))
 
-        assert len(remote_dataset.vector_settings) == 2
+        assert len(remote_dataset.vectors_settings) == 2
 
         remote_vector_settings = remote_dataset.vector_settings_by_name("other-vector")
         assert other_vector_settings.name == remote_vector_settings.name
         assert other_vector_settings.id == remote_vector_settings.id
         assert other_vector_settings.dimensions == remote_vector_settings.dimensions
 
-    def test_adding_vector_setting_with_the_same_name(self, owner: "User", feedback_dataset: FeedbackDataset):
+    def test_add_vector_setting_with_the_same_name(self, owner: "User"):
+        api.init(api_key=owner.api_key)
+        workspace = Workspace.create(name="test-workspace")
+
+        feedback_dataset = FeedbackDataset(fields=[TextField(name="text")], questions=[TextQuestion(name="text")])
+
+        remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
+        remote_dataset.add_vector_settings(VectorSettings(name="vector", dimensions=10))
+        dataset_vectors_settings = remote_dataset.vectors_settings
+        assert len(dataset_vectors_settings) == 1
+
+        with pytest.raises(ValueError, match="Vector settings with name 'vector' already exists"):
+            remote_dataset.add_vector_settings(VectorSettings(name="vector", dimensions=10))
+
+    def test_update_vectors_settings(self, owner: "User", feedback_dataset: FeedbackDataset):
+        api.init(api_key=owner.api_key)
+        workspace = Workspace.create(name="test-workspace")
+
+        feedback_dataset.add_vector_settings(VectorSettings(name="vector-settings-1", dimensions=4))
+        remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
+        vector_settings = remote_dataset.vector_settings_by_name("vector-settings-1")
+
+        vector_settings.title = "New Vector Settings Title"
+        remote_dataset.update_vectors_settings(vector_settings)
+
+        vector_settings = remote_dataset.vector_settings_by_name("vector-settings-1")
+        assert vector_settings.title == "New Vector Settings Title"
+
+    def test_bulk_update_vectors_settings(self, owner: "User", feedback_dataset: FeedbackDataset):
         api.init(api_key=owner.api_key)
         workspace = Workspace.create(name="test-workspace")
 
         remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
-        remote_dataset.add_vector_settings(VectorSettings(name="vector", dimensions=10))
-        dataset_vectors_settings = remote_dataset.vector_settings
-        assert len(dataset_vectors_settings) == 1
+        vectors_settings = list(remote_dataset.vectors_settings)
+        for vector_settings in vectors_settings:
+            vector_settings.title = "New Vector Settings Title"
 
-        with pytest.raises(ValueError, match=f"Vector settings with name 'vector' already exists"):
-            remote_dataset.add_vector_settings(VectorSettings(name="vector", dimensions=10))
+        remote_dataset.update_vectors_settings(vectors_settings)
 
-    def test_adding_records_with_vectors(self, owner: "User", feedback_dataset: FeedbackDataset):
+        for vector_settings in remote_dataset.vectors_settings:
+            assert vector_settings.title == "New Vector Settings Title"
+
+    def test_update_vectors_settings_one_by_one(self, owner: "User", feedback_dataset: FeedbackDataset):
+        api.init(api_key=owner.api_key)
+        workspace = Workspace.create(name="test-workspace")
+
+        remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
+
+        vectors_settings = list(remote_dataset.vectors_settings)
+        for vector_settings in vectors_settings:
+            vector_settings.title = "New Vector Settings Title"
+            remote_dataset.update_vectors_settings(vector_settings)
+
+        for vector_settings in remote_dataset.vectors_settings:
+            assert vector_settings.title == "New Vector Settings Title"
+
+    def test_delete_vectors_settings(self, owner: "User"):
+        api.init(api_key=owner.api_key)
+        workspace = Workspace.create(name="test-workspace")
+
+        feedback_dataset = FeedbackDataset(
+            fields=[TextField(name="text")],
+            questions=[TextQuestion(name="text")],
+            vectors_settings=[VectorSettings(name="vector", dimensions=4)],
+        )
+
+        remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
+        remote_dataset.delete_vectors_settings("vector")
+        assert len(remote_dataset.vectors_settings) == 0
+
+    def test_bulk_delete_vectors_settings(self, owner: "User"):
+        api.init(api_key=owner.api_key)
+        workspace = Workspace.create(name="test-workspace")
+
+        feedback_dataset = FeedbackDataset(
+            fields=[TextField(name="text")],
+            questions=[TextQuestion(name="text")],
+            vectors_settings=[
+                VectorSettings(name="vector-1", dimensions=4),
+                VectorSettings(name="vector-2", dimensions=4),
+                VectorSettings(name="vector-3", dimensions=4),
+            ],
+        )
+
+        remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
+
+        remote_dataset.delete_vectors_settings(["vector-1", "vector-2", "vector-3"])
+        assert len(remote_dataset.vectors_settings) == 0
+
+    def test_delete_vector_settings_one_by_one(self, owner: "User"):
+        api.init(api_key=owner.api_key)
+        workspace = Workspace.create(name="test-workspace")
+
+        feedback_dataset = FeedbackDataset(
+            fields=[TextField(name="text")],
+            questions=[TextQuestion(name="text")],
+            vectors_settings=[
+                VectorSettings(name="vector-1", dimensions=4),
+                VectorSettings(name="vector-2", dimensions=4),
+                VectorSettings(name="vector-3", dimensions=4),
+            ],
+        )
+
+        remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
+
+        for i, vector_settings in enumerate(remote_dataset.vectors_settings, 1):
+            deleted_vector_settings = remote_dataset.delete_vectors_settings(vector_settings.name)
+            assert deleted_vector_settings == vector_settings
+            assert len(remote_dataset.vectors_settings) == len(feedback_dataset.vectors_settings) - i
+
+        assert len(remote_dataset.vectors_settings) == 0
+
+    def test_add_records_with_vectors(self, owner: "User", feedback_dataset: FeedbackDataset):
         api.init(api_key=owner.api_key)
         workspace = Workspace.create(name="test-workspace")
 
@@ -492,19 +587,14 @@ class TestRemoteFeedbackDataset:
     @pytest.mark.parametrize(
         "invalid_vectors, expected_error",
         [
-            (
-                {"vector": [1, 1, 1, 1, 1]},
-                "Argilla server returned an error with http status: 422. "
-                "Error details: {'response': 'Record at position 0 is not valid because vector with name=vector is not valid",
-            ),
+            ({"vector": [1, 1, 1, 1, 1]}, "Vector with name `vector` has an invalid expected dimension."),
             (
                 {"unknown-vector": [1, 1, 1, 1]},
-                "Argilla server returned an error with http status: 422. "
-                "Error details: {'response': 'Record at position 0 is not valid because vector with name=unknown-vector is not valid",
+                "Vector with name `unknown-vector` not present on dataset vector settings.",
             ),
         ],
     )
-    def test_adding_records_with_invalid_vectors(
+    def test_add_records_with_invalid_vectors(
         self, owner: "User", feedback_dataset: FeedbackDataset, invalid_vectors: dict, expected_error: str
     ):
         api.init(api_key=owner.api_key)
@@ -514,7 +604,7 @@ class TestRemoteFeedbackDataset:
         feedback_dataset.add_vector_settings(VectorSettings(name="vector", dimensions=vector_dimension))
         remote_dataset = feedback_dataset.push_to_argilla(name="test_dataset", workspace=workspace)
 
-        with pytest.raises(ValidationApiError, match=expected_error):
+        with pytest.raises(ValueError, match=expected_error):
             remote_dataset.add_records(
                 FeedbackRecord(
                     fields={"text": "Hello world!", "text-2": "Hello world!"},