+++ tests/integration/client/sdk/v1/test_datasets.py
@@ -39,6 +39,7 @@ from argilla.client.sdk.v1.datasets.api import (
     get_records,
     list_datasets,
     publish_dataset,
+    search_records,
     update_records,
 )
 from argilla.client.sdk.v1.datasets.models import (
@@ -48,7 +49,10 @@ from argilla.client.sdk.v1.datasets.models import (
     FeedbackMetadataPropertyModel,
     FeedbackMetricsModel,
     FeedbackQuestionModel,
+    FeedbackRecordSearchModel,
     FeedbackRecordsModel,
+    FeedbackRecordsSearchModel,
+    FeedbackRecordsSearchVectorQuery,
     FeedbackSuggestionModel,
 )
 from argilla.client.sdk.v1.records.api import set_suggestion